+++ src/components/bluepad32/uni_platform_pc_debug.c
@@ -148,15 +148,18 @@ static int32_t pc_debug_get_property(uni_platform_property_t key) {
     return g_delete_keys;
 }
 
-static void pc_debug_on_device_oob_event(uni_hid_device_t* d, uni_platform_oob_event_t event) {
-    if (d == NULL) {
-        loge("ERROR: pc_debug_on_device_gamepad_event: Invalid NULL device\n");
-        return;
-    }
+static void pc_debug_on_oob_event(uni_platform_oob_event_t event, void* data) {
     logi("pc_debug: on_device_oob_event(): %d\n", event);
 
     if (event != UNI_PLATFORM_OOB_GAMEPAD_SYSTEM_BUTTON) {
-        loge("ERROR: pc_debug_on_device_gamepad_event: unsupported event: 0x%04x\n", event);
+        logi("pc_debug_on_device_gamepad_event: unsupported event: 0x%04x\n", event);
+        return;
+    }
+
+    uni_hid_device_t* d = data;
+
+    if (d == NULL) {
+        loge("ERROR: pc_debug_on_device_gamepad_event: Invalid NULL device\n");
         return;
     }
 
@@ -203,7 +206,7 @@ struct uni_platform* uni_platform_pc_debug_create(void) {
         .on_device_connected = pc_debug_on_device_connected,
         .on_device_disconnected = pc_debug_on_device_disconnected,
         .on_device_ready = pc_debug_on_device_ready,
-        .on_device_oob_event = pc_debug_on_device_oob_event,
+        .on_oob_event = pc_debug_on_oob_event,
         .on_gamepad_data = pc_debug_on_gamepad_data,
         .get_property = pc_debug_get_property,
     };